CPI-890 Clean up old snapshots only on successful snapshot store

+++ aggregate-snapshot/aggregate-snapshot-service/src/main/java/uk/gov/justice/services/eventsourcing/source/core/snapshot/DefaultSnapshotService.java
@@ -45,8 +45,10 @@ public class DefaultSnapshotService implements SnapshotService {
             try {
                 logger.trace("Storing snapshot of aggregate: {}, streamId: {}, version: {}", aggregate.getClass().getSimpleName(), streamId, streamVersionId);
                 final AggregateSnapshot<T> aggregateSnapshot = new AggregateSnapshot<>(streamId, streamVersionId, aggregate);
-                snapshotRepository.storeSnapshot(aggregateSnapshot);
-                snapshotRepository.removeAllSnapshotsOlderThan(aggregateSnapshot);
+                final boolean storedSuccessfully = snapshotRepository.storeSnapshot(aggregateSnapshot);
+                if(storedSuccessfully) {
+                    snapshotRepository.removeAllSnapshotsOlderThan(aggregateSnapshot);
+                }
             } catch (SerializationException e) {
                 logger.error("Error creating snapshot for {}", streamId, e);
             }
